Tests are now included and compiled
