+++ enumeratum-reactivemongo-bson/compat/src/main/scala-2.11/enumeratum/values/EnumHandler.scala
@@ -9,14 +9,11 @@ import scala.util.{ Failure, Success, Try }
  * @since 2016-04-23
  */
 object EnumHandler {
+
   /**
    * Returns a BSONReader for the provided ValueEnum based on the given base BSONReader for the Enum's value type
    */
-  def reader[ValueType <: AnyVal, EntryType <: ValueEnumEntry[ValueType]]
-    (enum: ValueEnum[ValueType, EntryType])
-    (implicit
-      baseBsonReader: BSONReader[BSONValue, ValueType]
-    ): BSONReader[BSONValue, EntryType] = new BSONReader[BSONValue, EntryType] {
+  def reader[ValueType <: AnyVal, EntryType <: ValueEnumEntry[ValueType]](enum: ValueEnum[ValueType, EntryType])(implicit baseBsonReader: BSONReader[BSONValue, ValueType]): BSONReader[BSONValue, EntryType] = new BSONReader[BSONValue, EntryType] {
     override def read(bson: BSONValue): EntryType = {
       val result: Try[EntryType] = baseBsonReader.readTry(bson).flatMap { s =>
         val maybeBound = enum.withValueOpt(s)
@@ -35,11 +32,7 @@ object EnumHandler {
   /**
    * Returns a BSONWriter for the provided ValueEnum based on the given base BSONWriter for the Enum's value type
    */
-  def writer[ValueType <: AnyVal, EntryType <: ValueEnumEntry[ValueType]]
-    (enum: ValueEnum[ValueType, EntryType])
-    (implicit
-      baseBsonWriter: BSONWriter[ValueType, BSONValue]
-    ): BSONWriter[EntryType, BSONValue] = new BSONWriter[EntryType, BSONValue] {
+  def writer[ValueType <: AnyVal, EntryType <: ValueEnumEntry[ValueType]](enum: ValueEnum[ValueType, EntryType])(implicit baseBsonWriter: BSONWriter[ValueType, BSONValue]): BSONWriter[EntryType, BSONValue] = new BSONWriter[EntryType, BSONValue] {
     override def write(t: EntryType): BSONValue = baseBsonWriter.write(t.value)
   }
 
@@ -47,14 +40,10 @@ object EnumHandler {
    * Returns a BSONHandler for the provided ValueEnum based on the given
    * base BSONReader and BSONWriter for the Enum's value type
    */
-  def handler[ValueType <: AnyVal, EntryType <: ValueEnumEntry[ValueType]]
-    (enum: ValueEnum[ValueType, EntryType])
-    (implicit
-       baseBsonReader: BSONReader[BSONValue, ValueType],
-       baseBsonWriter: BSONWriter[ValueType, BSONValue]
-    ): BSONHandler[BSONValue, EntryType] = new BSONHandler[BSONValue, EntryType] {
-    override def read(bson: BSONValue): EntryType = reader(enum).read(bson)
-
-    override def write(t: EntryType): BSONValue = writer(enum).write(t)
+  def handler[ValueType <: AnyVal, EntryType <: ValueEnumEntry[ValueType]](enum: ValueEnum[ValueType, EntryType])(implicit baseBsonHandler: BSONHandler[BSONValue, ValueType]): BSONHandler[BSONValue, EntryType] = new BSONHandler[BSONValue, EntryType] {
+    private val concreteReader = reader(enum)
+    private val concreteWriter = writer(enum)
+    override def read(bson: BSONValue): EntryType = concreteReader.read(bson)
+    override def write(t: EntryType): BSONValue = concreteWriter.write(t)
   }
 }
