+++ enumeratum-reactivemongo-bson/compat/src/main/scala-2.11/enumeratum/values/reactiveMongoBsonValueEnum.scala
@@ -1,5 +1,6 @@
 package enumeratum.values
 
+import EnumBSONHandlers._
 import reactivemongo.bson._
 
 /**
@@ -7,41 +8,8 @@ import reactivemongo.bson._
  * @since 2016-04-23
  */
 
-trait ReactiveMongoBsonValueEnum[ValueType <: AnyVal, EntryType <: ValueEnumEntry[ValueType]] {
+sealed trait ReactiveMongoBsonValueEnum[ValueType <: AnyVal, EntryType <: ValueEnumEntry[ValueType]] {
   enum: ValueEnum[ValueType, EntryType] =>
-
-  implicit val bsonReaderShort = new BSONReader[BSONValue, Short] {
-    override def read(bson: BSONValue): Short = bson match {
-      case BSONInteger(x) => x.toShort
-      case _ => throw new RuntimeException(s"Could not convert $bson to Short")
-    }
-  }
-
-  implicit val bsonReaderInt = new BSONReader[BSONValue, Int] {
-    override def read(bson: BSONValue): Int = bson match {
-      case BSONInteger(x) => x
-      case _ => throw new RuntimeException(s"Could not convert $bson to Int")
-    }
-  }
-
-  implicit val bsonReaderLong = new BSONReader[BSONValue, Long] {
-    override def read(bson: BSONValue): Long = bson match {
-      case BSONLong(x) => x
-      case _ => throw new RuntimeException(s"Could not convert $bson to Long")
-    }
-  }
-
-  implicit val bsonWriterShort = new BSONWriter[Short, BSONValue] {
-    override def write(t: Short): BSONValue = BSONInteger(t)
-  }
-
-  implicit val bsonWriterInt = new BSONWriter[Int, BSONValue] {
-    override def write(t: Int): BSONValue = BSONInteger(t)
-  }
-
-  implicit val bsonWriterLong = new BSONWriter[Long, BSONValue] {
-    override def write(t: Long): BSONValue = BSONLong(t)
-  }
 
   /**
    * Implicit BSON handler for the entries of this enum
