+++ enumeratum-reactivemongo-bson/compat/src/test/scala-2.11/enumeratum/values/BsonDrinks.scala
@@ -1,9 +1,9 @@
 package enumeratum.values
 
 /**
-  * @author Alessandro Lacava 
-  * @since 2016-04-23
-  */
+ * @author Alessandro Lacava
+ * @since 2016-04-23
+ */
 sealed abstract class BsonDrinks(val value: Short, name: String) extends ShortEnumEntry
 
 case object BsonDrinks extends ShortEnum[BsonDrinks] with ShortReactiveMongoBsonValueEnum[BsonDrinks] {
