+++ enumeratum-reactivemongo-bson/compat/src/test/scala-2.11/enumeratum/values/EnumBsonHandlerSpec.scala
@@ -1,10 +1,11 @@
 package enumeratum.values
 
 import org.scalatest._
+import EnumBSONHandlers._
 
 /**
-  * @author Alessandro Lacava (@lambdista)
-  * @since 2016-04-23
+ * @author Alessandro Lacava (@lambdista)
+ * @since 2016-04-23
  */
 class EnumBsonHandlerSpec extends FunSpec with Matchers with EnumBsonHandlerHelpers {
